Add files via upload
Overview:
WeatherCast is a user-friendly and intuitive web application designed to provide you with accurate and up-to-date weather information for any location in the world. Whether you're planning a trip, preparing for an outdoor activity, or simply want to stay informed about the weather in your area, WeatherCast has got you covered. With its sleek design and powerful features, the app offers a seamless experience that caters to both casual users and weather enthusiasts.

How WeatherCast Works:
WeatherCast leverages cutting-edge weather APIs to fetch accurate data from reputable weather sources. The data is then processed and presented in an easy-to-understand format on the web app. Users can interact with the app through an intuitive user interface that provides a seamless experience, whether accessing it from a browser or a mobile device.